requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0
xgboost>=1.5.0
lightgbm>=3.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
ta>=0.8.0
yfinance>=0.1.70
joblib>=1.1.0
pyyaml>=6.0
optuna>=2.10.0
jupyter>=1.0.0
notebook>=6.4.0